Nighthawk-PoC: state dump
- Move out the SSL stuff into separate files
- Use single pool interface abstraction for h2/h1
