+++ .vscode/c_cpp_properties.json
@@ -6,10 +6,8 @@
                 "/usr/include",
                 "/usr/local/include",
                 "${workspaceFolder}/bazel_output_base/external/**",
-                "${workspaceFolder}/envoy/source",
-                "${workspaceFolder}/envoy/include",
-                "${workspaceFolder}/source",
-                "${workspaceFolder}/include"
+                "${workspaceFolder}/envoy/**",
+                "${workspaceFolder}/**"
             ],
             "defines": [],
             "compilerPath": "/usr/bin/gcc",
@@ -21,10 +19,8 @@
                     "/usr/include",
                     "/usr/local/include",
                     "${workspaceFolder}/bazel_output_base/external/**",
-                    "${workspaceFolder}/envoy/source",
-                    "${workspaceFolder}/envoy/include",
-                    "${workspaceFolder}/source",
-                    "${workspaceFolder}/include"
+                    "${workspaceFolder}/envoy/**",
+                    "${workspaceFolder}/**"
                 ],
                 "limitSymbolsToIncludedHeaders": true,
                 "databaseFilename": ""
