+++ source/exe/client.h
@@ -15,8 +15,7 @@
 #include "common/runtime/runtime_impl.h"
 #include "common/thread_local/thread_local_impl.h"
 
-#include "common/http/http1/conn_pool.h"
-#include "common/http/http2/conn_pool.h"
+#include "envoy/http/conn_pool.h"
 #include "envoy/upstream/upstream.h"
 
 using namespace Envoy;
@@ -111,8 +110,7 @@ public:
                    Envoy::Upstream::HostDescriptionConstSharedPtr host) override;
 
 private:
-  std::unique_ptr<Envoy::Http::Http1::ConnPoolImpl> h1_pool_;
-  std::unique_ptr<Envoy::Http::Http2::ConnPoolImpl> h2_pool_;
+  Envoy::Http::ConnectionPool::InstancePtr pool_;
 };
 
 } // namespace Nighthawk
